Create main.yml
This commit adds a new file named 'main.yml'. This file contains the GitHub Actions workflow configuration for continuous integration (CI). The workflow is triggered on every push event to the repository. It includes a job named 'sast_scan', which runs the Bandit security scanning tool to analyze the code for potential security vulnerabilities. The workflow uses Python version 3.11.8 and installs the necessary dependencies before running the Bandit scan. This workflow will help ensure the security of the codebase by automatically scanning it for known security issues.